Renumber frames
For rationale, see https://docs.google.com/spreadsheets/d/1G73qp-GBhXGyKEsZBudzWJY85FSa5txT4MxtVC5ajiE/edit#gid=0

Note that this isn't 100% mechanical.  There are some other minor
changes:

Frames with multiple codepoints are not referred to in the plural in
their first mention (because we talk about multiple frame types), and
ahead of the figure that shows the fields.

I have made the text announcing the figure, and the text announcing the
fields consistent.

There was a grammatical error (the very last change here).