Refine minimum DCID length of Initial
This is meant to almost entirely mitigate #1810 by making sure that the reply to a VN packet ends up at the same server. It reduces the spurious downgrade attack signal to a corner case of a corner case.

It is much more lightweight than alternatives, if we're looking for a minimum-footprint change.