Use "Insert Count" rather than "Largest Reference"
The construction of largest reference currently implies the use of
1-based indexing, which is unnatural.  Using the insert count is much
more natural, even if it means the encoding is a little strange and a
tiny bit inefficient (it goes 0, 2, 3, ..., Max-1, 1, 2, ...).

I've gone through and put the references on the boundary between
records, which makes explanations much easier for post-base indexing in
my view.  I've renamed Base Index (which was never really an index) to
Base, which matches that change.  Similarly, renaming Largest Known
Received to Known Received Count makes it match better.

There are also examples added to Insert Count and Base sections.  These
definitions are pretty damned opaque without them.