Rework Key Update
This section was pretty old and it had at least one bug, along with a
bunch of editorial lack of clarity.

Substantively, this does two things:

1. It specifies what label to use for the KDF when updating.  The
agreement at one time was that we would use a quic-specific label, but
that never really got captured properly.  This proposes "quic ku".

2. It specifies that acknowledgments of packets with a given key phase
are necessary before another update can be triggered.  As noted in
 #2214, simultaneous updates can result in a deadlock if updates happen
too quickly.  This has the nice effect of allowing update compliance to
be more rigorously tested.

Closes #2214.