Use PTO to declare in-flight packets lost (#1965)
* Prevent an unlikely deadlock

QUIC, unlike TCP, allows the connection to get into a situation where there could be packets in flight, and a connection could even be CWND limited, but there may not be any outstanding data to retransmit.

The suggested solution is to send a retransmittable frame.

* Update draft-ietf-quic-recovery.md

* Update draft-ietf-quic-recovery.md

* Mike in-flight

* Update draft-ietf-quic-recovery.md

* Not a deadlock

* Update draft-ietf-quic-recovery.md

Co-Authored-By: ianswett <ianswett@users.noreply.github.com>

* Jana's suggestions

* Update draft-ietf-quic-recovery.md

* Proposed change: MUST mark inflight data as lost

A proposal: I think I prefer to mark things as lost instead of suggesting that a sender SHOULD send something when there's really nothing to be sent.  I was starting to rephrase your PR to say "SHOULD send ack-eliciting packet, but MAY mark as lost", and I needed to explain how a sender makes this choice. I realized that if it's been long enough (PTO is long enough) and there's nothing left to send, then maybe it is sensible to mark anything pending as lost. So, I changed my suggestion to that, which is this PR (against your PR).

* reflow

* add MAY re-arm PTO

* more rephrasing

* more simple

* comma