Get rid of DoS vulnerability in Reserved Bits
As currently written, any garbage packet, after removing header protection, will have nonzero reserved bits and bring the connection down. Not good!

Similarly, it is perverse to force me to actually decrypt the full packet to see if it turns out to be a misbehaving end host and comply with the other possible reading of this requirement. Let the receiver just drop the bad packet after header decryption and carry on.