Make key update optional
The important change here is to explicitly permit blocking of key
updates.  However, it does not remove the encouragement for enabling
them.  That seems to be the right sort of balance: we can't force
implementations to support updates, though it's still a good idea to do
so.

This makes more extensive changes to support better text in the
updated description of Key Update Permitted.  Formalizing "key phase" as
a concept rather than a field is the main one.