Expand the anti-replay story
The inclusion of RESET_STREAM (and all other frame types) in 0-RTT
inspired me to do some more analysis and due diligence on anti-replay.

This adds a security considerations section with requirements for
application protocols.  That section explains more about what the risks
for QUIC are, how QUIC itself isn't affected, but how an application
protocol might be.

It also adds a section to the HTTP draft citing RFC 8470, explaining how
that analysis is sufficient for HTTP/3.