Replay protection is the responsibility of application protocols
This is based on my recent conclusions about this subject.  It rewrites
the advice here by observing that QUIC does not inherently present a
replay risk. Instead, application protocols, in their use of QUIC, might
create an exposure to replay attack.

Rather than try to perform an analysis in the transport, based on
incomplete information, it is better to outline some risks (STREAM seems
like the only obvious one here, frankly, though I've pointed out a
couple of anti-patterns that might have accompanying replay risks) and
let the application protocol designers perform a more complete analysis.

We did that analysis for HTTP.  I believe that to be sufficient.  Though
I might include mention of the fact that stream cancellation and other
h2 mechanisms don't carry application semantics, but that is not a major
source of regret.