Make avoidable HTTP/3 errors connection errors
Fixes #2511

Does not include changes to the 4 causes I was unsure on below
"A PRIORITY frame sent on a request stream with the Prioritized Element Type set to any value other than "11" or which expresses a dependency on a request with a greater Stream ID than the current stream MUST be treated as a stream error of type HTTP_MALFORMED_FRAME."

In the appendix:
" SETTINGS_MAX_CONCURRENT_STREAMS: QUIC controls the largest open
Stream ID as part of its flow control logic. Specifying
SETTINGS_MAX_CONCURRENT_STREAMS in the SETTINGS frame is an error.

SETTINGS_INITIAL_WINDOW_SIZE: QUIC requires both stream and
connection flow control window sizes to be specified in the
initial transport handshake. Specifying
SETTINGS_INITIAL_WINDOW_SIZE in the SETTINGS frame is an error.

SETTINGS_MAX_FRAME_SIZE: This setting has no equivalent in HTTP/3.
Specifying it in the SETTINGS frame is an error."