Allow reuse of stateless reset tokens
This came up as the result of an incomplete implementation, but there
are - maybe - good reasons to have the same stateless reset token being
used for multiple connection IDs on the same connection.

There are drawbacks however.  Reuse of a stateless reset token means
that you have to remember (and consistently route) connection IDs, even
when they are retired.

This text makes a note of the possibility and explains the drawbacks.

Closes #2732.