Firm up language around stream termination
This was a little woolly.  Hopefully this is crisper.  Basically, the
application protocol is what asks for stream reset, either directly, or
by sending STOP_SENDING.

I added a recommendation for application protocols to define handling
for prematurely cancelled streams.  That is, I believe, just editorial,
even though it uses 2119 language, because we already do that.

Closes #2884.