Describe interaction between QUIC and TLS regarding saved 0-RTT state
This PR describes how a TLS stack needs to cooperate with a QUIC stack
when making the decision whether or not to accept early data.

The purpose of this change is so that implementors working on the TLS
layer of QUIC are aware that other layers impact decisions traditionally
made at the TLS layer.