Prevent linkability from responding to migration
This closes the rather serious hole we left when we attempted to limit
the effect of perpetual changing of connection IDs.  This uses a
narrower formulation, that I believe will avoid perpetual changes.  But
it does require reciprocal connection ID changes where endpoints
genuinely do migrate.

It's a design change unfortunately, but I hope non-controversial.

Closes #2778.