Perform stateless reset token comparisons in constant time
Otherwise information about the token might leak.

As @mikkelfj says, there is no strict need to compare across the entire
set of values.  That could leaks two things: that the inbound packet was
dropped and the total number of stateless reset tokens.  Both are things
that we might care about, but will probably find prohibitive to protect.

Closes #2152.