Suggest sending empty CRYPTO frames in Initial/Handshake packets
As per @tatsuhiro-t, since PING frames are not allowed in
Initial/Handshake packets, one can use an empty CRYPTO frame when
sending a probe when there is no other ack-eliciting frame to send.

It seems to me like this is something that is fairly easy to miss, so
worth expanding the existing suggestion I think.

The alternative would be to allow PING frames in Initial/Handshake, but
I figured we are maybe too late in the process for that kind of design
change.