Clarify the state a client stores with a token
It is permissible for a client to do 0-RTT resumption with a NST from
one connection and a token from another. This clarifies that a client
only needs to associate a token (from a NEW_TOKEN frame) with the server
it was from and no additional state, and that a server can't require
that a token be from the same connection as the NST in use.