Allow provisional registrations for all registries
This overhauls the system for IANA registries to use a unified scheme
that allows provisional registration.  This seems pretty complicated in
terms of nailing down all the corner cases, but it is fairly simple in
practice:

* Provisional registration is easy, it only requires that the requester
provide contact information and (ideally) the codepoint they want.
They could provide more, but at their discretion.

* Permanent registration requires that you fill out the fields.

Experts are required for both, but only really to prevent spam.  We have
a few more requirements for permanent registrations of frame types, but
that is more for the purposes of advice, because sending an new frame
leads to protocol errors if you haven't agreed to that in the past.

I've sketched out a strawman for a codepoint reclamation process.  That
advises that experts try to find some use of the codepoint.  There is a
date on registrations that can guide the selection of codepoints to
reclaim, but these can be updated any time by anyone who wants to keep
the registration "live".

Closes #3109. Closes #3020.