[HTTP/3] [editorial] 421 and future connection reuse.
Status code 421 does not indicate to clients that the connection should
not be reused for subsequent requests for a particular origin.  It only
indicates that the server is not able to produce a reponse to one
particular request.  See
https://httpwg.org/specs/rfc7540.html#MisdirectedRequest.

In case it is desirable that clients do not reuse the connection for
subsequent requests for a particular origin, that it a deviation from
the original description of 421 and should be called out explicitly.