Close in response to invalid CONNECTION_CLOSE
The looping case was a little unclear.  This should make it a tiny bit
clearer.  As Roni observes, we can't really let invalid CONNECTION_CLOSE
through without some sort of special handling.  This proposes that.

Unfortunately, this is a little more than an editorial fix.  So I'm
going to ask that the issue be promoted.

Closes #2475.