Remove some cruft from packet size section
This section had a bunch of text that wasn't updated as the result of
other changes elsewhere.  Though there was a suggestion that this could
reference other sections, but as this is in some ways the authoritative
section on the subject, I decided to keep the requirements.  That's
duplicative, but I couldn't see a simple way to deal with that.

Closes #3255.