Stateless immediate close
This says that immediate close when you don't have state doesn't have to
establish state.  Thus, if an endpoint has to immediately close, which
can do so without entering the closing period, which would establish
state.  It makes a general statement, which is that during the handshake
- that is, for unauthenticated inputs - packets can be discarded rather
than causing an immediate close.

This greatly simplifies the text for a small initial.  Now it only needs
to say "MUST discard, MAY immediate close as well".