Add handling of all packets marked as CE
This is actually a behaviour that has been observed a few time on the Internet. In TCP this case is encounter as ECN will be disabled if the SYN is CE marked (as the SYN should always be Not-ECT these days). I just realised that this validation is rather easy to detect if any Not-ECT have been sent, I think it would be good to mention is explicitly.