Sending probe packet has duplicated text
I removed this text because it was duplicated with text further below in the same section (but different normative requirements). I assume this was an oversight and the text further down is the more current one (at least it makes for sense to me). 

However, I'm not sure if this sentence is still needed, and if so why:
"In addition to sending data in the packet number space for which the timer
expired, the sender SHOULD send ack-eliciting packets from other packet
number spaces with in-flight data, coalescing packets if possible."
????