Text on session resumption
In looking at #3028, I realized that we had nowhere that addressed the
basic concept of resumption.  This is, I hope, all that we need to say
on the subject.  It talks about state and then the privacy implications
of using resumption.

I found less in the TLS 1.3 RFC on this subject than I might have liked
to see.  It only really addressed ticket reuse.  So this is a little
more verbose than is ideal.

Closes #3028.