Scrub for backticks
This doesn't do qpack because qpack is weird and needs a lot more work.

The places where I kept backticks are for the inline formulae that were
used for the reserved codepoints.  That seemed like it was worth quoting
somehow still.  I also kept backticks for references to variables in
examples.

However, this last principle really hurts in qpack and I don't know what
to do about that.

Closes #3609 (we'll need another issue for qpack).