Add a picture
This new figure shows the complete set of messages that determine the
connection ID that each peer uses.  The explanation describes how the
connection IDs in those messages are included in transport parameters.

Hopefully this aids in understanding how this is put together.  It's not
complicated, but it's not necessarily clear from reading a set of rules
that this is the case.