Exception for Retire Prior To
The rules for the active connection ID limit were such that the brief
bump that NEW_CONNECTION_ID + Retire Prior To could add was not
permitted.  Strictly speaking, if you requested retirement, the
connection IDs that you requested be retired are still active until you
receive a RETIRE_CONNECTION_ID frame.  That means that you were
technically violating this MUST in the spec.  Adding an exception for
this case was the best way I could see to fix this.

The other was to change the definition of "active connection ID".  I
tried that and, while it was a smaller change, it made the definition
less crisp.  This approach seems best.