Include max_udp_payload_size in packet size discussion
This isn't a complete dissertation on how clients might choose to use
larger packet sizes.  In particular, it doesn't capture the whole
discussion about starting at a higher size being a way to raise the 1280
to something higher, with the requisite cost being that you might have
failures.  Deciding to put that in would be much harder.

What this does is just say that the transport parameter acts as a
further limit on packet size and that you won't know that prior to
knowing.  That means that you are guessing about the network path AND
your peer when you decide to send >1200 (to be fair, the former really
only applies for IPv4, and it's not much of a guess).

Closes #3638.