Second set of editorial nits
@gorryfair had some useful feedback in #3735.

This doesn't address the following items as numbered in the issue:

(2) Because the fix to (1) seems adequate by my reckoning.  That
addresses the question of considering the impact on other network users.

(6) This doesn't appear to be a problem.  It might be a formatting
issue, with the URL not appearing.

(7) This is addressed in the recent draft by adding [Section
10.2.2](https://quicwg.org/base-drafts/draft-ietf-quic-transport.html#name-deferring-idle-timeout)

(9) and (11) I can't find the rendering glitch mentioned.

The rest of these are good.

Closes #3735.