Make language in ECN validation clearer
This avoids "no smaller than" and other such ambiguities.

However, in doing this, I realized that I might interpret the original
language as allowing a network to ADD markings.  That seems wrong.  But
it's a non-editorial change.  This version has that change.  I'll make
another PR with the hole left in place.

Closes #3733.