Rework ECN validation rules
This attempts to create rules that won't break down unnecessarily.  It's
a little overwrought though, so I'm OK if we go with something less than
this.  The new check number 3 could be removed if we were comfortable
with the simplified check number 2 and the addition of the check for
unused markings.

Closes #3778.