Clear vestiges of path validation symmetry
In early versions of path validation, there was a requirement that
PATH_CHALLENGE be sent on the same path.  This is, for various reasons,
either unnecessary or harmful, so we no longer have that requirement.
But there was some residual text that assumed that property.

This removes a paragraph citing symmetry directly, and rewords text that
assumes return reachability - a property that QUIC no longer provides
mechanisms for.

Closes #3797.