Reorder TLS interface summary
I came back to this today and found that the ordering was inconsistent
and not sufficiently clear.

Changing this so that all the actions for a single endpoint were grouped
seemed to be the best arrangement.  All events and actions are as close
as possible in the ordering to the preceding event that caused or
enabled it.

The application data packets are displayed with a different marking,
which helps distinguish those that carry CRYPTO frames.

The final implied exchange for Get Handshake was removed.  This was, I
think, intended to imply something about session tickets.  Without
explanation, it is confusing.

This doesn't show HANDSHAKE_DONE, nor do I think that it should, because
that is something that the transport does once it gets the Handshake
Complete signal.