More about authenticating Version Negotiation
This clarifies that it is a mechanism in the target QUIC version that
authenticates the information.

This also lifts the strong requirement on authenticating the contents of
Version Negotiation.  QUIC version 1 does not do this and so does not
comply with this requirement.  The intent was always that only
information that is used needs to be authenticated in this way.

(This is a tiny bit of evasion as Version Negotiation is used in QUIC
version 1 as a signal that a version is not supported, which results in
clients abandoning connections.)

Closes #3828.