Be clearer about protections
This breaks down some of the packet protection text in favour of more
specific language about the protections provided.  By preference,
this talks about integrity protection and confidentiality protection as
it relates to those packets that are protected using keys derived from
TLS.  For Initial packets this means avoid use of "packet protection" as
much as possible (I'm sure that I have missed some of these).  The same
for Retry.

This required some additional framing, but it wasn't too intrusive.
Most of the changes are cosmetic.

Closes #3887.