Application can reset stream or close connection even if no error occurs.
The application may reset a stream, for example, when a browser tab is
closed or when the pushed resource is found in cache.  It might close
the connection when the application is closed.  These are not errors per
se, even though the term "stream error" and "connection error" is used
for these events, and they have an associated "error code".