Strike confusing paragraph
This text could be read to imply that an off-path attacker is more
capable than an on-path attacker, which is rarely true.  What it was
meant to point out was that it is easier to move traffic onto a path
that you are on.  What it fails to acknowledge is that it is also easier
to move traffic *off* a path that you are on.

In other words, the treatment of this in 21.12 is more thorough and we
don't need to talk about limitations.

Mike suggested that there is some duplication between this attack and
the more comprehensive analysis in 21.12.  That is true, but these serve
different purposes.  This is to describe attacks and the normative
requirements on endpoints necessary to avoid them.  The other section is
a thorough and hollistic analysis.  I couldn't see any truly
straightforward changes.  That doesn't mean that we won't find a way to
clean this up, or that it would be undesirable to have fewer words, but
I've not the time for that right now.

Closes #3841.