More words on the minimum packet size requirement
For stateless resets, we recommend a minimum packet size on ALL packets,
based on the maximum length of connection IDs that might be generated by
the same endpoint.  This isn't exactly clear, but nor do I think we need
to include the fully worked logic that leads to it.

This attempts to split the difference.

Closes #3848.