Closing and draining tidy
I took a careful look at the text here and found a few things:

1. There wasn't much redundancy, but there was a little.

2. The closing and draining states only apply to immediate close.

So I moved things up.  Shortened the top-level immediate close text, and
moved text specific to closing or draining to the corresponding
sections.

There were a few bits that could be cut, and a few that needed to be
better aligned with existing text, but nothing major.  Most of this is
unmodified text.

There are a few bits that I had to rewrite.  I will add comments for
those.

Closes #3907.