Application can reset stream or close connection even if no error occurs. (#3953)
* Application can reset stream or close connection even if no error occurs.

The application may reset a stream, for example, when a browser tab is
closed or when the pushed resource is found in cache.  It might close
the connection when the application is closed.  These are not errors per
se, even though the term "stream error" and "connection error" is used
for these events, and they have an associated "error code".

* Update draft-ietf-quic-http.md

Co-authored-by: Mike Bishop <mbishop@evequefou.be>

* Jana's suggestion to pluralize errors.

* Jana's suggestion to pluralize even more

Co-authored-by: Jana Iyengar <jri.ietf@gmail.com>

* Mike's wording of "indicate a problem"

Co-authored-by: Mike Bishop <mbishop@evequefou.be>

Co-authored-by: Bence Béky <bnc@google.com>
Co-authored-by: Mike Bishop <mbishop@evequefou.be>
Co-authored-by: Jana Iyengar <jri.ietf@gmail.com>