Document request forgery
First cut at this, sorry about it being so long.  It's fairly hard to
write this, particularly since the existing documentation is so sparse.
Hopefully this is clear enough.  I haven't had time to properly
proof-read it, so it's probably not very good.  But I thought I'd share
what I got.

This includes some basic countermeasures, but they aren't very good.
I think that's OK, but we should discuss.

Closes #3995.