Out of order relative to other ack-eliciting packets
I believe this was the existing intent, so not filing an issue, people can feel free to disagree.

The case I'm concerned about is an ACK arrives with PN 10, then PN 9 arrives.  I do not believe PN 9 should not be considered 'out-of-order' and trigger an immediate ACK.  Recovery is designed with this case in mind, but when re-reading this text, I thought it was unclear.