Persistent congestion pseudocode to match text
This adds a new variable `first_rtt_sample` that records when the first
RTT was recorded.  This variable is used to avoid declaring persistent
congestion based on packets that were sent prior to this time.

Note that I removed the lame check in `InPersistentCongestion`.  This
avoids the implication that you need to propagate a signal that says
"this is the first RTT sample" all the way from the ACK handling code.
That signal is now propagated using the new variable.

Closes #3972.