Describe inputs to TLS more clearly
This clarifies the TLS interface by explaining that TLS only generates
outputs (handshake bytes or keying material) in response to inputs.
Previously, the only inputs we acknowledged were the handshake bytes,
but this recognizes two more: the signal from the client to start
handshaking, and - for async validation - a signal that a certificate is
OK or not.

Closes #3873.