Fix the AllPacketsLost thing too
This turned out to be really tricky.  There is still a "cheat" here in a
function I called `FindLargestContiguousLoss`, which requires comments.

I did manage to address the problem of doing this across packet number
spaces by introducing a global variable.  That variable now collects all
lost packets into a single place.

However, doing that means that you need to clear it out occasionally, or
you declare persistent congestion all the time.  This only clears it out
each time you declare persistent congestion, which is massively
inefficient.  I hope people realize that, but it didn't seem
appropriate as the best fix I could come up with requires breaking down
the clean separation of congestion control and loss recovery parts.

(That fix is to reset this persistent congestion collection if the
largest acknowledged increases, but after running the persistent
congestion check.  That requires that the loss recovery code
and congestion control code be interleaved in awkward ways.  Maybe there
is a simpler approach, but I can't think of one right now.)

cc @ianswett, @janaiyengar