WIP: Begin lightly abstracting over the use of UDP as the underlying transport
Based on recent mailing list feedback that this might possibly be OK, I
decided to take a crack at reducing the "hard coding" of UDP in the
spirit of other RFCs which strive to be agnostic to the underlying
protocol.

Due to the controversy over this, I imposed these constraints on myself
to try to be as conservative as possible.

- Absolutely no design changes on things which are already standardized,
  just editorial changes, and very simple extrapolation of the existing
  design. (I never ever considered the former at this time, just
  documenting this for posterity.

- Anything that is not UDP (over IP) is deemed experimental. The size
  limitations I figure may not make sense for other transports / during
  other experiments, so I call that out.

- When UDP was mentioned in some interesting way that generalized (e.g.
  something that separately mentions IP address and UDP port), I provide
  the general language (plain "address", but also provide the UDP common
  case as before so no specificity is lost.

The purpose of this exercise is *not* to get QUIC over UDP "more ready
for production", as that would slow down the WG for little gain, but
rather to call out the dependencies that currently do exist so as to be
sure there is no unintentional coupling. I fully leave to others to
decide what coupling is intenional or unintentional, and whether
anything at all is cause for concern.

----

In this first commit, I just audited all the occurrence of "UDP". If
this looks good, I would then go back and likewise scan for "port"
"IPv4", and "IPv6".

My one regret is that in keeping the line length the same, I made the
diff more complicated than it would otherwise be. Do you all have a
process for that?