Pseudocode for persistent congestion is too hard
This removes that function, relying instead on comments and the
definition in text (which people have noted is now very clear).

I have also reset a couple of extra variables in this case.  It is
important to catch all of these because if you don't you get poor
outcomes.

The only question I have is regarding bytes in flight.  I think that we
need to keep tracking those bytes, or the overall accounting gets messy.
That is what our implementation does anyway.

Closes #4010.
Closes #3972.